requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==3.2.16
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
